src/loader/requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-documentai
functions-framework
